Chore - refactor util files for separation of concerns
