+++ plugins/toolkit/jetbrains-core/src/software/aws/toolkits/jetbrains/services/codemodernizer/state/CodeModernizerState.kt
@@ -9,7 +9,7 @@ import com.intellij.openapi.projectRoots.JavaSdkVersion
 import com.intellij.util.xmlb.annotations.Property
 import software.aws.toolkits.jetbrains.services.codemodernizer.model.CodeModernizerSessionContext
 import software.aws.toolkits.jetbrains.services.codemodernizer.model.JobId
-import software.aws.toolkits.jetbrains.services.codemodernizer.toVirtualFile
+import software.aws.toolkits.jetbrains.services.codemodernizer.utils.toVirtualFile
 
 enum class JobDetails {
     LAST_JOB_ID,
@@ -48,7 +48,7 @@ class CodeModernizerState : BaseState() {
     fun getLatestJobId() = JobId(lastJobContext[JobDetails.LAST_JOB_ID] ?: throw RuntimeException("No Job has been executed!"))
 
     fun toSessionContext(project: Project): CodeModernizerSessionContext {
-        val configurationFile = lastJobContext[JobDetails.CONFIGURATION_FILE_PATH]?.toVirtualFile()
+        val configurationFile = software.aws.toolkits.jetbrains.services.codemodernizer.utils.toVirtualFile()
             ?: throw RuntimeException("No build file store in the state")
         val targetString =
             lastJobContext[JobDetails.TARGET_JAVA_VERSION] ?: throw RuntimeException("Expected target language for migration path of previous job but was null")
