+++ plugins/toolkit/jetbrains-core/src/software/aws/toolkits/jetbrains/services/codemodernizer/utils/ModuleUtils.kt
@@ -1,0 +1,20 @@
+// Copyright 2024 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+package software.aws.toolkits.jetbrains.services.codemodernizer.utils
+
+import com.intellij.openapi.module.Module
+import com.intellij.openapi.project.Project
+import com.intellij.openapi.projectRoots.JavaSdkVersion
+import com.intellij.openapi.projectRoots.impl.JavaSdkImpl
+import com.intellij.openapi.roots.ModuleRootManager
+import com.intellij.openapi.roots.ProjectRootManager
+
+/**
+ * @description Try to get the module SDK version and fallback to the project SDK version from the "project structure" settings.
+ */
+fun Module.tryGetJdk(project: Project): JavaSdkVersion? {
+    val sdk = ModuleRootManager.getInstance(this).sdk ?: ProjectRootManager.getInstance(project).projectSdk ?: return null
+    val javaSdk = JavaSdkImpl.getInstance()
+    return javaSdk.getVersion(sdk)
+}
